facilitate intersect function
Even though the previous one worked, replace it for a simpler one.

This works because two ranges that does not overlap when the end of
one range comes before the start of the other. I.e. a1 < b0 or a0 > b1.

What we search for in the intersect function is the opposite:
NOT (a1 < b0 OR a0 > b1).

Using Demorgan's Law this yields the expression:
a1 >= b0 AND a0 <= b1